pyproject.toml
[project]
name = "luminalink"
version = "0.1.0"
description = "Blind-friendly movie narration system (vision-to-text, phase 1)"
readme = "README.md"
authors = [
    { name = "Bryan Kuang", email = "2455350756@qq.com" }
]
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.4.1",
    "opencv-python>=4.13.0.90",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "setuptools>=80.9.0",
    "typer>=0.21.1",
    "webrtcvad>=2.0.10",
]

[project.scripts]
luminalink = "luminalink:main"

[project.optional-dependencies]
vision = [
    "facenet-pytorch>=2.5.3",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]
nlg = [
    "tensorflow>=2.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
